9 aug 11 pm

+++ controllers/expenseControllers.js
@@ -38,11 +38,15 @@ const getExpenses = async (req, res) => {
 
 const deleteExpense = async (req, res) => {
     try {
-        const expense = await Expense.findByIdAndDelete(req.params.id);
-        if(!expense){
-            return res.status(404).json({msg:"Required expense not found"})
+        const expense = await Expense.findById(req.params.id);
+        if (!expense) {
+            return res.status(404).json({ msg: "Required expense not found" })
         }
-        res.status(200).json({msg:"Expense deleted successfully"})
+        if (expense.userId != req.user) {
+            return res.status(403).json({ msg: "Access denied" })
+        }
+        await Expense.findByIdAndDelete(req.params.id);
+        res.status(200).json({ msg: "Expense deleted successfully" })
     } catch (err) {
         console.log("Error:", err);
         res.status(500).json({ msg: "Internal server error" });
